delete with wab page works fine
but when user recommend comment, and delete it cause
MySQLIntegrityConstraintViolationException
becasue comment_id is foreign key in table comment, recommended_comment

need to set table cascade

+++ src/main/java/kr/ac/jejunu/service/CommentService.java
@@ -47,7 +47,7 @@ public class CommentService {
 
 	public void deleteComment(String id, int commentId) {
 		Comment comment = commentDao.getCommentById(commentId);
-		if (comment.getWriter().getId().equals(id))
+		if (comment != null && comment.getWriter().getId().equals(id))
 			commentDao.delete(commentId);
 		else
 			throw new OwnerNotMatchedException();
